Feature: remove blocking delay and now is waaaaaaaay faster. Feature: use uint16_t as channel values Fix: use uint32_t as header

+++ include/constants.h
@@ -16,7 +16,7 @@ namespace fortress {
     }
 
     namespace net {
-        enum MsgTypes : uint8_t {
+        enum MsgTypes : uint32_t {
             ServerAccept,
             ServerDeny,
             ServerMessage,
